// evio_Identity_Microservice/utils/enums/requestEmailChange.ts
export enum requestEmailChange {
    // Redis
    redisKey = "CHANGE_EMAIL_REQUEST",
    billingChangeEmail = "billing_profile_change_email",
    userChangeEmail = "user_change_email",
    // Email Templates
    welcomeEmailTemplate = "welcome",
    billingCreateEmailTemplate = "create_billing_profile"
}